+++ Runtime/OpenXR/EyeTracker/MagicLeapEyeTrackerNativeFunctions.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 using UnityEngine.XR.OpenXR.NativeTypes;
 
@@ -11,7 +12,8 @@ namespace MagicLeap.OpenXR.Features.EyeTracker
         internal delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerGeometricDataGetInfo, uint, out uint, XrEyeTrackerGeometricData*, XrResult> XrGetEyeTrackerGeometricData;
         internal delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerPupilDataGetInfo, uint, out uint, XrEyeTrackerPupilData*, XrResult> XrGetEyeTrackerPupilData;
         internal delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerGazeBehaviorGetInfo, ref XrEyeTrackerGazeBehavior, XrResult> XrGetEyeTrackerGazeBehavior;
-
+        internal delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerPosesGetInfo, ref XrEyeTrackerPoses, XrResult> XrGetEyeTrackerPoses;
+        
         protected override void Validate()
         {
             base.Validate();
@@ -44,6 +46,11 @@ namespace MagicLeap.OpenXR.Features.EyeTracker
             {
                 Debug.LogError($"Unable to find {nameof(XrGetEyeTrackerGazeBehavior)}");
             }
+            
+            if (XrGetEyeTrackerPoses == null)
+            {
+                Debug.LogError($"Unable to find {nameof(XrGetEyeTrackerPoses)}");
+            }
         }
 
         private string SanitizeFunctionName(string input)
@@ -59,6 +66,7 @@ namespace MagicLeap.OpenXR.Features.EyeTracker
             XrGetEyeTrackerGeometricData = (delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerGeometricDataGetInfo, uint, out uint, XrEyeTrackerGeometricData*, XrResult>)LocateNativeFunction(SanitizeFunctionName(nameof(XrGetEyeTrackerGeometricData)));
             XrGetEyeTrackerPupilData = (delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerPupilDataGetInfo, uint, out uint, XrEyeTrackerPupilData*, XrResult>)LocateNativeFunction(SanitizeFunctionName(nameof(XrGetEyeTrackerPupilData)));
             XrGetEyeTrackerGazeBehavior = (delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerGazeBehaviorGetInfo, ref XrEyeTrackerGazeBehavior, XrResult>)LocateNativeFunction(SanitizeFunctionName(nameof(XrGetEyeTrackerGazeBehavior)));
+            XrGetEyeTrackerPoses = (delegate* unmanaged[Cdecl]<ulong, in XrEyeTrackerPosesGetInfo, ref XrEyeTrackerPoses, XrResult>)LocateNativeFunction(SanitizeFunctionName(nameof(XrGetEyeTrackerPoses)));
         }
     }
 }