+++ Runtime/Subsystems/Input/Extensions/Utils.cs
@@ -1,7 +1,7 @@
 // %BANNER_BEGIN%
 // ---------------------------------------------------------------------
 // %COPYRIGHT_BEGIN%
-// Copyright (c) (2019-2022) Magic Leap, Inc. All Rights Reserved.
+// Copyright (c) (2019-2024) Magic Leap, Inc. All Rights Reserved.
 // Use of this file is governed by the Software License Agreement, located here: https://www.magicleap.com/software-license-agreement-ml2
 // Terms and conditions applicable to third-party materials accompanying this distribution may also be found in the top-level NOTICE file appearing herein.
 // %COPYRIGHT_END%
@@ -18,9 +18,13 @@ namespace UnityEngine.XR.MagicLeap
         public static class Utils
         {
             /// <summary>
-            /// Returns valid Magic Leap input device for provided <seealso cref="InputDeviceCharacteristics"/>.
-            /// If there is none matching device the returned <seealso cref="InputDevice.isValid"/> gonna be set as false.  
+            /// Retrieves a Magic Leap input device that matches the specified <see cref="InputDeviceCharacteristics"/>.
             /// </summary>
+            /// <param name="inputDeviceCharacteristics">The characteristics to match against input devices.</param>
+            /// <returns>
+            /// A valid <see cref="InputDevice"/> that matches the specified characteristics.
+            /// If no matching device is found, an invalid <see cref="InputDevice"/> is returned (i.e., <see cref="InputDevice.isValid"/> will be <c>false</c>).
+            /// </returns>
             public static InputDevice FindMagicLeapDevice(InputDeviceCharacteristics inputDeviceCharacteristics)
             {
                 List<InputDevice> devices = new List<InputDevice>();
@@ -45,9 +49,10 @@ namespace UnityEngine.XR.MagicLeap
             /// </param>
             /// <param name="predictSnapshots">Boolean representing whether or not to predict the snapshot.</param>
             public static void PredictSnapshot(long timestamp, bool predictSnapshots) => MagicLeapXrProviderNativeBindings.PredictSnapshot(timestamp, predictSnapshots);
-            
+
             /// <summary>
-            /// Reset the snapshot state of the subsystem to before prediction. Use this to cleanup after PredictSnapshot
+            /// Resets the snapshot prediction state of the input subsystem. 
+            /// Call this method to clean up after using <see cref="PredictSnapshot"/>.
             /// </summary>
             public static void ResetSnapshotPrediction() => MagicLeapXrProviderNativeBindings.ResetSnapshotPrediction();
         }